+++ Content.Client/UserInterface/Systems/Chat/ChatUIController.cs
@@ -928,7 +928,7 @@ public sealed partial class ChatUIController : UIController
     private void OnChatMessage(MsgChatMessage message)
     {
         var msg = message.Message;
-        ProcessChatMessage(msg, !msg.HidePopup);
+        ProcessChatMessage(msg, !msg.HidePopup); // Trauma added msg.HidePopup
 
         if ((msg.Channel & ChatChannel.AdminRelated) == 0 ||
             _config.GetCVar(CCVars.ReplayRecordAdminChat))
