+++ Resources/Locale/en-US/_Trauma/phones/phones.ftl
@@ -19,3 +19,15 @@ phone-category-science = science
 phone-category-service = service
 phone-category-uncategorized = uncategorized
 phone-category-hidden = hidden
+
+signal-port-name-phone-call = On call recieved
+signal-port-description-phone-call = Called whenever the phone starts to ring.
+
+signal-port-name-phone-outgoing = On outgoing call
+signal-port-description-phone-outgoing = Called whenever someone starts an outgoing call.
+
+signal-port-name-phone-pickup = On phone pickup
+signal-port-description-phone-pickup = Called whenever someone picks up the phone.
+
+signal-port-name-phone-hangup = On phone hangup
+signal-port-description-phone-hangup = Called whenever someone hands up the phone.
