+++ Resources/Prototypes/_Goobstation/Catalog/Fills/Crates/service.yml
@@ -313,16 +313,3 @@
     - id: RotaryPhoneFlatPack
     - id: RotaryPhoneFlatPack
     - id: RotaryPhoneFlatPack
-
-- type: entity
-  id: CrateHolySupplies
-  parent: CratePlastic
-  name: holy supplies crate
-  description: Contains a Bible and 2 Jugs of holy water for religious needs
-  components:
-    - type: StorageFill
-      contents:
-        - id: JugHolyWater
-          amount: 2
-        - id: Bible
-          amount: 1
