+++ Resources/Prototypes/_Goobstation/Entities/Objects/Devices/flatpack.yml
@@ -242,14 +242,3 @@
   components:
   - type: Flatpack
     entity: Vaccinator
-
-- type: entity
-  parent: BaseFlatpack
-  id: RotaryPhoneFlatPack
-  name: phone flat pack
-  description: A flatpack used for constructing a Rotary Phone.
-  components:
-  - type: Flatpack
-    entity: RotaryPhoneStand
-  - type: Item
-    size: Normal
