+++ Resources/Prototypes/_Goobstation/Entities/Objects/Devices/pda.yml
@@ -366,15 +366,6 @@
     id: PhoneMerchantIDCard
 
 - type: entity
-  parent: BasePDA
-  id: PhoneMerchantPDA
-  name: phone merchant PDA
-  description: Ring a ding ding!
-  components:
-  - type: Pda
-    id: PhoneMerchantIDCard
-
-- type: entity
   parent: CentcomPDA
   id: OuterCommandPDA
   name: outer command PDA
