+++ Resources/Prototypes/_Goobstation/Entities/Objects/Misc/identification_cards.yml
@@ -385,18 +385,6 @@
   - type: PresetIdCard
     job: PhoneMerchant
 
-- type: entity
-  parent: IDCardStandard
-  id: PhoneMerchantIDCard
-  name: phone merchant ID card
-  components:
-  - type: Sprite
-    layers:
-    - state: default
-    - state: idphonemerchant
-  - type: PresetIdCard
-    job: PhoneMerchant
-
 # Outer Command
 - type: entity
   parent: CentcomIDCardNavyCaptain
