+++ Resources/Prototypes/_Goobstation/Roles/play_time_trackers.yml
@@ -79,6 +79,3 @@
 
 - type: playTimeTracker
   id: JobPhoneMerchant
-  
-- type: playTimeTracker
-  id: JobOuterCommander
