+++ Resources/Prototypes/_Goobstation/StatusIcon/job.yml
@@ -52,11 +52,3 @@
     sprite: /Textures/_Goobstation/Interface/Misc/job_icons.rsi
     state: SecurityClown
   allowSelection: false
-
-- type: jobIcon
-  parent: JobIcon
-  id: JobIconPhoneMerchant
-  icon:
-    sprite: /Textures/_Goobstation/Interface/Misc/job_icons.rsi
-    state: phonemerchant
-  jobName: job-name-phonemerchant
