+++ Resources/Prototypes/_Goobstation/Voice/speech_sounds.yml
@@ -25,12 +25,3 @@
     path: /Audio/Items/brokenbikehorn.ogg
   exclaimSound:
     path: /Audio/Items/desk_bell_ring.ogg
-
-- type: speechSounds
-  id: RMCPhoneSpeakSounds
-  saySound:
-    collection: RMCPhoneSpeak
-  askSound:
-    collection: RMCPhoneSpeak
-  exclaimSound:
-    collection: RMCPhoneSpeak
