+++ Resources/Prototypes/_Lavaland/Entities/Objects/Misc/identification_cards.yml
@@ -8,4 +8,16 @@
     - state: default
     - state: idshaftminer
   - type: PresetIdCard
-    job: ShaftMiner
+    job: ShaftMiner  
+
+- type: entity
+  parent: IDCardStandard
+  id: PhoneMerchantIDCard
+  name: phone merchant ID card
+  components:
+  - type: Sprite
+    layers:
+    - state: default
+    - state: idphonemerchant
+  - type: PresetIdCard
+    job: PhoneMerchant
