+++ Resources/Prototypes/_Trauma/Catalog/Cargo/service.yml
@@ -1,0 +1,9 @@
+- type: cargoProduct
+  id: RotaryPhoneCrateCargo
+  icon:
+    sprite: _RMC14/Objects/phone/rotary_phone.rsi
+    state: rotary_phone
+  product: CrateRotaryPhone
+  cost: 2500
+  category: cargoproduct-category-name-service
+  group: market
