+++ Resources/Prototypes/_Trauma/Catalog/Fills/Crates/service.yml
@@ -1,0 +1,13 @@
+- type: entity
+  id: CrateRotaryPhone
+  parent: CrateGenericSteel
+  name: rotary phone crate
+  description: A crate with 5 phones inside, prank calls are a punishable offense
+  components:
+  - type: StorageFill
+    contents:
+    - id: RotaryPhoneFlatPack
+    - id: RotaryPhoneFlatPack
+    - id: RotaryPhoneFlatPack
+    - id: RotaryPhoneFlatPack
+    - id: RotaryPhoneFlatPack
