+++ Resources/Prototypes/_Trauma/Entities/Objects/Devices/flatpack.yml
@@ -1,0 +1,10 @@
+- type: entity
+  parent: BaseFlatpack
+  id: RotaryPhoneFlatPack
+  name: phone flat pack
+  description: A flatpack used for constructing a Rotary Phone.
+  components:
+  - type: Flatpack
+    entity: RotaryPhoneStand
+  - type: Item
+    size: Normal
