+++ Resources/Prototypes/_Trauma/Entities/Objects/Devices/phones.yml
