+++ Resources/Prototypes/_Trauma/StatusIcon/job.yml
@@ -1,0 +1,8 @@
+- type: jobIcon
+  parent: JobIcon
+  id: JobIconPhoneMerchant
+  icon:
+    sprite: /Textures/_Goobstation/Interface/Misc/job_icons.rsi
+    state: phonemerchant
+  jobName: job-name-phonemerchant
+  allowSelection: false
