+++ Resources/Prototypes/_Wega/Roles/play_time_trackers.yml
@@ -1,2 +1,5 @@
 - type: playTimeTracker
   id: JobGeneticist
+
+- type: playTimeTracker
+  id: JobPhoneMerchant
